Create unique_paths2.py
This problem determines the number of unique paths possible when there are obstacles in the 2d matrix with 1 being an obstacle and 0 being free space. The goal is to reach the bottom right corner from the top left corner of the matrix, while you can only move either rightside or downwards from your current cell.