Create wallsandgates.py
This problem finds the shortest distance of each empty cell from a gate consider walls that don't allow a person to pass, the wall is represented as -1, the empty space is represented as INF and the empty space as 0. We find each shortest distance from their gates and place that value in that particular cell.